+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.2.0
+- improvement: add explaining labels to the dialog modes https://github.com/jardicc/alchemist/issues/8
 - improvement: option to choose font size: https://github.com/jardicc/alchemist/issues/5
 - improvement: adds possibility to show message strip with link
 - fix: wrong layer property names https://github.com/jardicc/alchemist/issues/13 https://github.com/jardicc/alchemist/issues/11